# src/docs/openapi/afenda-public.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Afenda documented HTTP API",
    "version": "1.0.0",
    "description": "Machine-readable index for API routes documented by the Afenda evidence manifests. Generated API evidence remains the ownership source; this OpenAPI document gives Fumadocs a canonical API rendering surface."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the web application"
    }
  ],
  "paths": {
...
    "/api/tenants/{tenant}/admin/integrations/conflicts": {
      "get": {
        "tags": ["Integration Sync"],
        "operationId": "integration-sync.conflicts",
        "summary": "Return integration sync conflicts.",
        "description": "See generated evidence: /docs/generated/api/integration-sync.conflicts",
        "parameters": [{ "$ref": "#/components/parameters/Tenant" }],
        "responses": {
          "200": {
            "description": "Request completed successfully"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Tenant": {
        "name": "tenant",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Tenant slug."
      },
...
    }
  }
}